// src/store/questions/types.ts
export enum Actions {
    GET_ITEMS = 'GET_ITEMS',
    ADD_ITEM = 'ADD_ITEM',
    UPDATE_ITEM = 'UPDATE_ITEM',
    DELETE_ITEM = 'DELETE_ITEM',
}

export enum Getters {
    ITEM_LIST = 'ITEM_LIST',
}

export enum Mutations {
    RESET = 'RESET',
    SET_ITEMS = 'SET_ITEMS',
    ADD_ITEM = 'ADD_ITEM',
    REPLACE_ITEM = 'REPLACE_ITEM',
    DELETE_ITEM = 'DELETE_ITEM',
}